PPhotonic_Radar_AI_Project-master/requirements.txt
# Primary Dependencies
numpy
scipy
matplotlib
streamlit
torch
torchvision
torchaudio
opencv-python
pyrtlsdr
kafka-python
psutil
seaborn
scikit-learn
plotly

# Configuration & logging helpers
pyyaml
python-json-logger

# Optional / Legacy
# tensorflow (Not supported on Python 3.14 yet)
